pyproject.toml
[project]
name = "saturn-mousehunter-strategy-engine"
version = "0.1.0"
description = "Saturn MouseHunter Strategy Engine - Comprehensive strategy management and backtesting service"
readme = "README.md"
authors = [
    { name = "cenwei", email = "cenwei@msn.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyJWT>=2.8.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.2"
]

[project.scripts]
saturn-mousehunter-strategy-engine = "saturn_mousehunter_strategy_engine:main"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"
